Create Learning-Numpy
From what i know so far, numpy arrays takes significantly less amount of memory compared to python lists. It also provides a mechanism of specifying the data types of the contents, which allows further optimisation of the code. An example of this can include a small amount of numbers stored in an array and because the data is so small, it doesn't need a set data size of 32 bytes to store it.