Add files via upload
Small project using the numpy module. Created surface of a 7x7 numpy array, all values are set to zero except for the the number in the middle of the surface which is a number 1, you can move this number 1 around this array-plane, however, if you go beyond the 7x7 area, you have fell out of the world.