Add files via upload
Using dictionary password to crack the hashed passwords using the 'bcrypt' module.

This project was something I have done while studying, however, processing the hashes using the bcrypt module can be quite slow, in addition i had this initial issue of the script only going through one iteration of passwords for one user name before ending the script. 

To fix these issues, I first split up the script into functions to make things more manageable. I thought that this could fix the looping problem because the script is now reduced from a complicated program with a lot of nested loops, now into more smaller chunks, which reduces the overall complexity of what the program has to process. 

The other issue of it not processing the hashes and passwords fast enough to my liking was a tough thing to solve, I came across two methods such as using multi-threading, that utilises the ability of the processor to execute multiple threads concurrently, hence why I split the functions into two halves, both working on cracking.the script hashes concurrently. The other solution is using caching it allow the script to cache the return values of a function depending on the arguments.

Challenge:
A text file contains user names and Unix hashed passwords from a Linux server. The user names and hashed passwords are separated by a white space.
Your task is to write a program which:
    Read user names and hashed passwords  (file attached)
    Crack hashed passwords using the dictionary password file
    Create a new file name correct_passwords.txt which includes user names and crack passwords (separate by space) if password is found.
    User names and hashed passwords